code-refactor - splited large components into small components by passing props
