+++ src/client/components/clinicPortal/patients/barcode/index.jsx
