+++ src/client/components/clinicPortal/patients/barcode/scanBarComponent.jsx
