+++ src/client/components/clinicPortal/patients/barcode/scanBarResult.jsx
