+++ src/client/components/clinicPortal/patients/qrScanReader/index.jsx
