+++ src/client/services/common/errorHandler.js
@@ -1,0 +1,6 @@
+export const handleError = (error) => {
+    console.log(error) // 401
+    if(error.status === 401){
+        window.location.href = "/home";
+    }
+};
